(#002) Update README.md
Creation content of the README file

This commit creates the README file with a unique template for the JosephGallegos repository. This template is released under a Creative Commons license. If you want to use it, please do not change the authorship of this README file.

This README uses http://img.shields.io to include badges, declaring the use of different technologies.

Signed-off-by: Joseph Gallegos <jaremgallegosp@gmail.com>